+++ CHANGELOG.md
@@ -5,6 +5,11 @@ This project adheres to [Semantic Versioning](http://semver.org/).
 
 ## Unreleased
 
+### Added
+
+* New process type `notify`
+  ([#41](https://gitlab.com/veenj/cinit/issues/41))
+
 ## [1.3.9]
 
 ### Added